Make load_data module; make target shape optional; fix line endings.
